Added files for project
Automaticdai project spawned with intent to create an application specific to loading a generic cart with objects. The algorithm can detect one main reference object according to a combination of colors, then detect N-number of objects that are within the reference object according to additional color combinations. The algorithm returns a bounding box around the ROI's for purposes of downstream automation.